Add Thunderbird customisation and refactor folder structuring for Firefox and Thunderbird customisation files
Introduced Thunderbird customisation files. In this commit, the
customisation focuses on mail previewer window, where it inverts the
dark theme's background, font and button colors to make it more distinguishable.
These changes will hamper if the light theme is used. Hence to be used
only in case of dark theme. A future scope might be to look into a way
identify the theme current applied.

Restructured the Firefox and Thunderbird customisation files to avoid
overlap as the file names are the same.

The files are to be used as soft symlinks to the actual profiles and would
make it easier for modifying the changes in the repo location would
reflect in the application profile as well.